+++ R/kwic_method.R
@@ -1,7 +1,7 @@
 #' @include partition_class.R context_class.R
 NULL 
 
-#' KWIC output / concordances
+#' KWIC/concordance output.
 #' 
 #' Prepare and show concordances / keyword-in-context (kwic). The same result can be achieved by 
 #' applying the kwic method on either a partition or a context object.
@@ -27,6 +27,11 @@ NULL
 #' @rdname kwic
 #' @docType methods
 #' @seealso To read the whole text, see the \code{\link{read}}-method.
+#' @references 
+#' Baker, Paul (2006): \emph{Using Corpora in Discourse Analysis}. London: continuum, pp. 71-93 (ch. 4).
+#'
+#' Jockers, Matthew L. (2014): \emph{Text Analysis with R for Students of Literature}.
+#' Cham et al: Springer, pp. 73-87 (chs. 8 & 9).
 #' @examples
 #' \dontrun{
 #' if (require(polmineR.sampleCorpus) && require(rcqp)){